feat: Started implementing the bottom navbar
