+++ src/layouts/AppLayout.tsx
@@ -1,4 +1,6 @@
-import LeftNavbar from "@/components/NavBars/LeftNavbar.tsx";
+import BottomNavBar from "@/components/NavBars/BottomNavBar.tsx";
+import LeftNavBar from "@/components/NavBars/LeftNavBar";
+import { NavBarLocation } from "@/types/payloads/ready.ts";
 import { useSettingsStore } from "@/wrapper/Stores.ts";
 import { memo } from "react";
 
@@ -7,7 +9,7 @@ const AppLayout = memo(({ children }: { children?: React.ReactElement | React.Re
 
 	return (
 		<div className="flex">
-			{navBarLocation === "left" && <LeftNavbar />}
+			{navBarLocation === NavBarLocation.Left ? <LeftNavBar /> : <BottomNavBar />}
 			{children}
 		</div>
 	);
